Restructured to include gcc-4.9 toolchain built libs Added to toolchain to into build control [ndk-build friendly names]

+++ utils/build_shaderc.sh
@@ -14,6 +14,17 @@
 # Usage:                                                                 #
 #     modify the $TARGET_ROOT, $SRC_ROOT, $SCRIPT_ROOT                   #
 #     then run this script                                               #
+# Generated File Structure                                               #
+#       shaderc                                                          #
+#           toolChain-type (build only one toolchain at a run )          #
+#               stl-type1 [total types are configurable]                 #
+#                   arch1                                                #
+#                   arch2                                                #
+#                   arch3                                                #
+#               stl-type2                                                #
+#                   arch1                                                #
+#                   arch2                                                #
+#                   arch3                                                #
 # Tested:                                                                #
 #     Mac OS                                                             #
 ##########################################################################
@@ -27,32 +38,40 @@ AR_TOOL=/Users/gfan/dev/ndk_current/toolchains/x86_64-4.9/prebuilt/darwin-x86_64
 
 ## shaderc source tree
 SRC_ROOT=${HOME}/backup/shaderc/src
+## should pull source code?
+#PULL_SRC=YES
+
+# Toolchain to use[4.8, 4.9, clang3.5, clang3.6]
+toolChain=4.9
+TOOLCHAIN=NDK_TOOLCHAIN_VERSION=$toolChain
 
 ## location to host combined libs
-TARGET_ROOT=${HOME}/backup/shaderc/result/shaderc
+TARGET_ROOT=${HOME}/backup/shaderc/result/shaderc/${toolChain}
 COMBINED_LIB_NAME=libshaderc.a
 
 ## STL_TYPE
 STL_TYPES=(c++_static gnustl_static)
 
 ## arch types we are trying to build for given stl_type
-declare ARCHS=(x86 x86_64 armeabi armeabi-v7a arm64-v8a mips mips64)
+declare ARCHS=(x86 x86_64 armeabi armeabi-v7a arm64-v8a)
 
 #### pull source code into SRC_ROOT####################
-rm  -fr ${SRC_ROOT}
-git clone https://github.com/google/shaderc ${SRC_ROOT}
-pushd $SRC_ROOT/third_party
-git clone https://github.com/google/googletest.git
+if [ -n "${PULL_SRC}" ] && [ "${PULL_SRC}" == "YES" ]; then
+  rm  -fr ${SRC_ROOT}
+  git clone https://github.com/google/shaderc ${SRC_ROOT}
+  pushd $SRC_ROOT/third_party
+  git clone https://github.com/google/googletest.git
 
-#  after vulkan is released, get from github
-#    https://github.com/google/glslang.git
-#    https://github.com/KhronosGroup/SPIRV-Tools.git spirv-tools
-git clone git@gitlab.khronos.org:GLSL/glslang.git
-cd glslang
-git checkout -b vulkan-glsl origin/KHR_vulkan_glsl
-cd ..
-git clone git@gitlab.khronos.org:spirv/spirv-tools.git
-popd
+  #  after vulkan is released, get from github
+  #    https://github.com/google/glslang.git
+  #    https://github.com/KhronosGroup/SPIRV-Tools.git spirv-tools
+  git clone git@gitlab.khronos.org:GLSL/glslang.git
+  cd glslang
+  git checkout -b vulkan-glsl origin/KHR_vulkan_glsl
+  cd ..
+  git clone git@gitlab.khronos.org:spirv/spirv-tools.git
+  popd
+fi
 
 #####build and combine libs ###########################
 if [ ! -f ${SCRIPT_ROOT}/combine.mri ]; then
@@ -66,6 +85,8 @@ for stl in "${STL_TYPES[@]}"; do
 
   pushd ${SRC_ROOT}/android_test
   ndk-build clean
+  grep -q 'NDK_TOOLCHAIN_VERSION' ./jni/Application.mk  && sed -i .org "s/.*NDK_TOOLCHAIN_VERSION.*/$TOOLCHAIN/g" ./jni/Application.mk || echo "${TOOLCHAIN}" >> ./jni/Application.mk
+
   sed -i .org ${SUB_CMD} ./jni/Application.mk
   ndk-build
